Bench: Restore the memory harness --max-retained-growth flag
Camel-casing the flag names gave `maxRetainedGrowth`, but the schema key is
`maxRetainedGrowthMb`, so the value never landed and Zod fell back to the 400MB
default. parseArgs accepted the flag, nothing warned, and the harness compared
against a threshold nobody asked for. It needs the same explicit rename as
--out and --json.

args.test.ts now carries a flag with the same shape, so removing the rename
fails the suite instead of going quiet.

Also finish the react-renderer-module rename, which left generate-project.ts
importing a deleted file, and correct the SaveSample comment: saves count
from 1, not 0.
