+++ scripts/bench/docgen-shared/budgets.ts
@@ -1,8 +1,3 @@
-/**
- * Memory budget values for the engine gates. Budgets are derived per engine from that engine's own
- * baseline runs and are never ported between engines; missing engines get their rows when their
- * baselines are recorded.
- */
 import type { EngineId } from './engine-ids.ts';
 
 export interface MemoryBudgets {
@@ -14,18 +9,10 @@ export interface MemoryBudgets {
   maxRetainedSlopeMb: number;
 }
 
-/**
- * Budgets sit well above observed values so the gate is not flaky, while still failing hard on a
- * real regression.
- */
-export const MEMORY_BUDGETS: Partial<Record<EngineId, MemoryBudgets>> = {
+const MEMORY_BUDGETS: Partial<Record<EngineId, MemoryBudgets>> = {
   'react-osa': { maxRetainedGrowthMb: 60, maxTransientMb: 90, maxRetainedSlopeMb: 3 },
 };
 
-/**
- * Budgets for `engine`, or a hard failure. A gate that reads a missing row would assert nothing and
- * still pass, so an engine without recorded budgets must not reach a gate config at all.
- */
 export function memoryBudgetsFor(engine: EngineId): MemoryBudgets {
   const budgets = MEMORY_BUDGETS[engine];
   if (!budgets) {
