+++ scripts/bench/docgen-shared/child-output.ts
@@ -1,7 +1,3 @@
-/**
- * The last `count` lines of a measured child's combined stdout and stderr, indented as a log block.
- * Enough to see why a run failed without spilling a full V8 crash dump into the output.
- */
 export function outputTail(output: string, count: number): string {
   return output
     .trim()
