+++ scripts/bench/docgen-shared/engine-ids.ts
@@ -1,6 +1,5 @@
 /**
- * The docgen engines under measurement. Lives here rather than with the perf suite's result types
- * because the memory gate keys its budgets off the same ids.
+ * The docgen engines under measurement.
  */
 export type EngineId =
   | 'react-legacy'
